Initial implementation of Core Data was added
The data is now saved in core data so it will stay there until it is deleted although I haven't removed the file or code that originally saved the data, I only commented out the code that saved the data using that file. I also created "groups" and organized some of the files into those groups to make it easier to read the files rather than a bunch of files in a long list.