ScrollView enlarged and added a toolbar
The toolbar is for the Y/N textfield's pickerView and right now it only has a done button. Also I made a small change so that only the first 4 items in the subviews add the button for images so a spacing issue between those that don't have that button for images came up.