Fixed the problem when trying to save edited information
I fixed the issue of creating an entirely new inspection form when it was supposed to edit an existing inspection form. I also added a scrollview to the tableCellViewController so the user can scroll through the entire inspection form. I also added a breakpoint to try to catch the problem when trying to edit the 2nd, 3rd, 4th, or 5th image that crashes the program.