Fixed the program crashing upon trying to edit any button except the 1st one
I fixed the issue where the program would crash when the user tried to edit the image for any button except the 1st one. I also found numerous bugs relating to how the buttons are being updated, saved into coreData, and viewed when more than 1 inspection category has more than 1 button. The program now crashes when it tries to get all of the images when saving any new edits into coreData because it walks out of bounds because it never sees a "nil" or default button because when the user edits the last button and there is less than 5 or so buttons in the category there needs to be a new button created that is the "nil" or default button.