Reverted recently made changes to the ibhandler
The changes that were made to the ibhandler didn't help the issue it instead created more problems and confusion so they were changed back to what it was originally. Although there were some improvements regarding the button IDs now having unique IDs so when the didSelect function is called in the InspectionFormViewController file it knows exactly which button was selected so we can easily find which button inside the ibhandler needs to be changed.