Made changes to the order of the categories
I made changes to the order of the categories based off of John's word document, I also enabled buttons to be available for every category and I added a button in the InspectionFormViewController that will increase the height of the scrollView. That function will need to be implemented differently for when the button goes onto the next row to shift everything down and to make sure the user can view the entire inspection form.