Commented out print Picture String
When I ran the program and published an inspection form with only 1 picture when printing out the Picture String (line 310 in Database file) I would see a long line of text before the line of text just disappears and nothing else get printed after it. This for whatever reason also prevents the 2 print statements above it to not appear in the output either.